% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFunc.R
\name{pFunc}
\alias{pFunc}
\title{A parametric g formula prediction function}
\usage{
pFunc(mod, ndat)
}
\arguments{
\item{mod}{A model object to predict from.}

\item{ndat}{The new data to predict from.}
}
\description{
This function is used by the pgf function to predict
values from model objects.
}
\examples{
pFunc()
}
\keyword{g-formula}
