% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgf_v0.0.0.9000.R
\name{pgf}
\alias{pgf}
\title{A parametric g formula prediction function}
\usage{
pgf(ii, mc_data, length, randomization = NULL, exposure = NULL,
  censoring = NULL)
}
\arguments{
\item{ii}{A numeric sequence (e.g., 1:5000) indicating monte
carlo resample size.}

\item{mc_data}{The original re-sampled baseline data.}

\item{length}{A numeric length of follow-up to be simulated.}

\item{randomization}{A numeric randomization. "NULL" indicates
natural course, "1" indicates set randomization to treatment,
"0" indicates set randomization to placebo. Defaults to "NULL".}

\item{exposure}{A numeric exposure. "NULL" indicates natural course,
"1" indicates set to exposed, "0" indicates set to unexposed.}

\item{censoring}{A numeric censoring. "NULL" indicates no
censoring, and "natural" indicates censoring as in empirical
data. Defaults to "NULL".}
}
\description{
This function allows you simulate the follow-up of a complex
longitudinal dataset under different intervention regimes.
}
\examples{
pgf()
}
\keyword{causal}
\keyword{g-formula,}
\keyword{inference}
