% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_function.R
\name{SL.model}
\alias{SL.model}
\title{A SuperLearner Function}
\usage{
SL.model(dependent, independent, id, library, cores)
}
\arguments{
\item{dependent}{A numeric dependent variable, binomial or gaussian.}

\item{independent}{A matrix with all independent variables.}

\item{id}{A numeric individual identifier.}

\item{library}{A list of SL libraries.}
}
\description{
This function allows you to fit the SuperLearner.
}
\examples{
SL.model()
}
\keyword{SuperLearner}
