% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_function.R
\name{SL.model}
\alias{SL.model}
\title{A SuperLearner Function}
\usage{
SL.model(multicore = 0, dependent, independent, id, library)
}
\arguments{
\item{multicore}{A binary integer (0,1) variable to choose parallel
or sequential processing. If "0", standard SuperLearner is
used. If "1" mcSupearLearner is used. Defaults to "0".}

\item{dependent}{A numeric dependent variable, binomial or gaussian.}

\item{independent}{A matrix with all independent variables.}

\item{id}{A numeric individual identifier.}

\item{library}{A list of SL libraries.}
}
\description{
This function allows you to fit the SuperLearner.
}
\examples{
SL.model()
}
\keyword{SuperLearner}
